Fix binding after `canBindAs` call
When `canBindAs` is called the binding changes from the initial binding that
has been configured. Therefore follow up communication (such as requesting
groups and user info) may fail due to missing permissions of the bound context.

HINT: This is just a quickfix. It would be better to use a dedicated
connection within `canBindAs`, so the initial connection is not messed up.

Change-Id: I418e68558f36141ac24d2b5b91b213b3b04089a7