Circumvent apparent LDAP binding for unknown users
If LDAP is configured to let anyone bind, then for any username a log in was attempted, as the search for the username returned an empty string. As a result canBindAs returned true.
This lead to an Error:
' PluggableAuthPrimaryAuthenticationProvider returned an invalid username '

With this commit canBindAs returns false on such empty username.

Cherrypicking to REL1_31 recommended.

Change-Id: I49e0fca62aad0038d973c14efc62f50f63183153