Add primary user group
In my previous patch, I was only looking at the groups a user is member
of but lacking the primary group a user belongs to.

Now the group specified by the users Group-ID is also added to the list
of user groups.

This is an improved version which solves the issues with accessing now
protected variables.

Signed-off-by: Philipp Kolmann <philipp@kolmann.at>
Change-Id: I620dbf8bf4f545b8567c61761b89dedad73c0d76